+++ lib/src/websocket/websocket.dart
@@ -1,6 +1,4 @@
 export 'websocket_event_types.dart';
-export 'websocket_events.dart';
 export 'websocket_state.dart';
 export 'websocket_stats.dart';
 export 'websocket_stream.dart';
-export 'websocket_stream_transformer.dart';